Move game components into $lib/ui/game
