+++ src/lib/ui/game/BoardView.svelte
