+++ src/lib/ui/game/Sidebar.svelte
