+++ src/lib/ui/game/board-selection/BoardPreview.svelte
