+++ src/lib/ui/game/board/BattleMap.svelte
