+++ src/lib/ui/game/board/BoardTools.svelte
