+++ src/lib/ui/game/board/Panel.svelte
