+++ src/lib/ui/game/board/grid/GridPattern.svelte
