+++ src/lib/ui/game/board/grid/SvgHexagon.svelte
