+++ src/lib/ui/game/board/token-palette/TokenTemplateItem.svelte
