+++ src/lib/ui/game/board/token-properties/Inheritable.svelte
