+++ src/lib/ui/game/board/token-properties/TokenPropertiesPanel.svelte
