+++ src/lib/ui/game/board/token-properties/TokenPropertyAvatar.svelte
