+++ src/lib/ui/home/CampaignDeleteDialog.svelte
