+++ src/lib/ui/home/forms/ActivationCodeForm.svelte
