+++ src/packages/ui/layout/PanView.svelte
